+++ src/components/setting/appearances/SkinSearchbar.tsx
@@ -1,81 +1,48 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
-import React, { useState } from 'react';
-import { Searchbar, ProgressBar } from 'react-native-paper';
-import { View, StyleSheet } from 'react-native';
+import React from 'react';
 import { useTranslation } from 'react-i18next';
 
 import { useNoxSetting } from '@stores/useApp';
-import useSnack from '@stores/useSnack';
 import logger from '@utils/Logger';
+import SearchBar from '@components/commonui/SearchBar';
+import { getUniqObjects } from '@utils/Utils';
 
 interface Props {
-  onSearched: (val: any) => void;
+  getThemeID: (skin: NoxTheme.Style) => string;
 }
-const CustomSkinSearch = ({
-  onSearched = (vals: any) => console.log(vals),
-}: Props) => {
-  const { t } = useTranslation();
-  const setSnack = useSnack(state => state.setSnack);
-  const [searchVal, setSearchVal] = useState(
-    'https://raw.githubusercontent.com/lovegaoshi/azusa-player-mobile/master/src/components/styles/steria.json',
-  );
-  const [searchProgress, progressEmitter] = useState(0);
-  const playerStyle = useNoxSetting(state => state.playerStyle);
 
-  const handleSearch = async (val = searchVal) => {
-    progressEmitter(1);
-    try {
-      const res = await fetch(val);
-      const searchedResult = await res.json();
-      onSearched(searchedResult);
-    } catch (e) {
-      logger.warn(`[SkinSearchbar] failed to search ${e}`);
-      setSnack({
-        snackMsg: { success: t('CustomSkin.SearchFailMsg') },
-      });
-    } finally {
-      progressEmitter(0);
+export default ({ getThemeID }: Props) => {
+  const { t } = useTranslation();
+  const playerStyles = useNoxSetting(state => state.playerStyles);
+  const setPlayerStyles = useNoxSetting(state => state.setPlayerStyles);
+  const loadCustomSkin = (skins: NoxTheme.Style[]) => {
+    // skins MUST BE an array of objects
+    if (!Array.isArray(skins)) {
+      throw new Error('requested skin URL is not an array. aborting.');
     }
+    const uniqueSkins = getUniqObjects(
+      skins.filter(skin => skin.metaData).concat(playerStyles),
+      getThemeID,
+    );
+    setPlayerStyles(uniqueSkins);
   };
 
   return (
-    <View style={styles.container}>
-      <View style={styles.row}>
-        <Searchbar
-          placeholder={t('CustomSkin.SearchBarLabel')}
-          value={searchVal}
-          onChangeText={setSearchVal}
-          onSubmitEditing={() => handleSearch(searchVal)}
-          selectTextOnFocus
-          style={styles.textInput}
-          selectionColor={playerStyle.customColors.textInputSelectionColor}
-          onIconPress={() => handleSearch(searchVal)}
-          numberOfLines={1}
-        />
-      </View>
-      <ProgressBar
-        progress={Math.max(searchProgress, 0)}
-        indeterminate={searchProgress === 1}
-        style={styles.progressBar}
-      />
-    </View>
+    <SearchBar
+      defaultSearchText="https://raw.githubusercontent.com/lovegaoshi/azusa-player-mobile/master/src/components/styles/steria.json"
+      onSearch={async ({ v, setSnack }) => {
+        try {
+          const res = await fetch(v);
+          const searchedResult = await res.json();
+          loadCustomSkin(searchedResult);
+        } catch (e) {
+          logger.warn(`[SkinSearchbar] failed to search ${e}`);
+          setSnack({
+            snackMsg: { success: t('CustomSkin.SearchFailMsg') },
+          });
+        }
+      }}
+      placeholder={t('CustomSkin.SearchBarLabel')}
+    />
   );
 };
-
-const styles = StyleSheet.create({
-  container: {
-    width: '100%',
-    paddingHorizontal: 5,
-    paddingTop: 5,
-  },
-  row: {
-    flexDirection: 'row',
-    width: '100%',
-  },
-  textInput: {
-    flex: 5,
-  },
-  progressBar: { backgroundColor: 'rgba(0, 0, 0, 0)' },
-});
-
-export default CustomSkinSearch;
